#41 add @ParametrizedTest annotation to tests
