+++ pom.xml
@@ -319,6 +319,12 @@ SOFTWARE.
       <artifactId>commons-io</artifactId>
       <version>2.11.0</version>
     </dependency>
+      <dependency>
+          <groupId>org.junit.jupiter</groupId>
+          <artifactId>junit-jupiter-params</artifactId>
+          <version>5.9.3</version>
+          <scope>test</scope>
+      </dependency>
   </dependencies>
   <profiles>
     <profile>
