+++ src/test/kotlin/org/objectionary/deog/unit/graph/attr/AttrTest.kt
@@ -24,48 +24,75 @@
 
 package org.objectionary.deog.unit.graph.attr
 
-import org.junit.jupiter.api.Test
+import org.objectionary.deog.graph.GraphBuilder
+import org.objectionary.deog.graph.repr.DGraphNode
+import org.objectionary.deog.sources.SrsTransformed
+import org.objectionary.deog.sources.XslTransformer
+import org.objectionary.deog.steps.AttributesSetter
+import org.objectionary.deog.unit.graph.TestBase
+import org.apache.commons.io.FileUtils
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.CsvSource
+import org.slf4j.LoggerFactory
+import java.io.BufferedReader
+import java.io.ByteArrayOutputStream
+import java.io.File
+import java.nio.file.Path
+import java.nio.file.Paths
 
 /**
- * Tests attributes propagation
+ * Base class for attributes propagation testing
  */
-class AttrTest : AttrBase() {
-    @Test
-    fun `test basic tree`() = doTest()
+open class AttrTest : TestBase {
+    private val logger = LoggerFactory.getLogger(this.javaClass.name)
+    private var postfix = "tmp"
 
-    @Test
-    fun `test multiple aliases`() = doTest()
+    @ParameterizedTest
+    @CsvSource(value = [
+        "basic_cycle",
+        "basic_dir",
+        "basic_tree",
+        "closed_cycle",
+        "inner",
+        "inner_concrete",
+        "inner_ordered",
+        "multiple_aliases",
+        "multiple_closed_cycles",
+        "multiple_cycles",
+        "multiple_trees",
+        "tree",
+        "triple_cycle"
+    ], ignoreLeadingAndTrailingWhitespace = true)
+    fun doTest(testName: String) {
+        val sources = SrsTransformed(Path.of(constructInPath(testName)), XslTransformer(), postfix)
+        val graph = GraphBuilder(sources.walk()).createGraph()
+        val attributesSetter = AttributesSetter(graph)
+        attributesSetter.setAttributes()
+        val out = ByteArrayOutputStream()
+        printOut(out, graph.dgNodes)
+        val actual = String(out.toByteArray())
+        val bufferedReader: BufferedReader = File(constructOutPath(testName)).bufferedReader()
+        val expected = bufferedReader.use { it.readText() }
+        checkOutput(expected, actual, "In test: ${Path.of(constructInPath(testName))}")
+        try {
+            val tmpDir =
+                Paths.get("${constructInPath(testName)}_$postfix").toString()
+            FileUtils.deleteDirectory(File(tmpDir))
+        } catch (e: Exception) {
+            logger.error(e.printStackTrace().toString())
+        }
+    }
 
-    @Test
-    fun `test tree`() = doTest()
+    override fun constructOutPath(directoryName: String): String =
+        "src${sep}test${sep}resources${sep}unit${sep}out${sep}attr$sep$directoryName.txt"
 
-    @Test
-    fun `test multiple trees`() = doTest()
-
-    @Test
-    fun `test basic cycle`() = doTest()
-
-    @Test
-    fun `test triple cycle`() = doTest()
-
-    @Test
-    fun `test multiple cycles`() = doTest()
-
-    @Test
-    fun `test closed cycle`() = doTest()
-
-    @Test
-    fun `test multiple closed cycles`() = doTest()
-
-    @Test
-    fun `test inner`() = doTest()
-
-    @Test
-    fun `test inner concrete`() = doTest()
-
-    @Test
-    fun `test inner ordered`() = doTest()
-
-    @Test
-    fun `test basic dir`() = doTest()
+    private fun printOut(
+        out: ByteArrayOutputStream,
+        nodes: Set<DGraphNode>
+    ) {
+        nodes.sortedBy { it.name }.forEach { node ->
+            out.write("NODE: ${node.name} ATTRIBUTES:\n".toByteArray())
+            node.attributes.forEach { out.write("name=${it.name}, dist=${it.parentDistance}\n".toByteArray()) }
+        }
+    }
 }
