+++ src/test/kotlin/org/objectionary/deog/unit/graph/sources/tempdir/TempDirectoryBase.kt
@@ -3,6 +3,8 @@ package org.objectionary.deog.unit.graph.sources.tempdir
 import org.objectionary.deog.sources.SrsTransformed
 import org.objectionary.deog.sources.XslTransformer
 import org.objectionary.deog.unit.graph.TestBase
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.CsvSource
 import java.io.File
 import java.nio.file.Path
 import kotlin.test.assertTrue
@@ -12,7 +14,31 @@ import kotlin.test.assertTrue
  */
 open class TempDirectoryBase : TestBase {
     private val postfix = "tmp"
-    override fun doTest() {
+
+    /**
+     * Tests different directories formats as input
+     *
+     * @param testName test name
+     */
+    @ParameterizedTest
+    @CsvSource(
+        value = [
+            "basic_cycle",
+            "basic_dir",
+            "basic_tree",
+            "closed_cycle",
+            "inner",
+            "inner_concrete",
+            "inner_ordered",
+            "multiple_aliases",
+            "multiple_closed_cycles",
+            "multiple_cycles",
+            "multiple_trees",
+            "tree",
+            "triple_cycle"
+        ], ignoreLeadingAndTrailingWhitespace = true
+    )
+    fun doTest(testName: String) {
         for (i in 0..2) {
             val path = constructInPath(getTestName()) + sep.toString().repeat(i)
             deleteTempDirIfExists(Path.of(path))
@@ -35,5 +61,6 @@ open class TempDirectoryBase : TestBase {
         val tempDir = File(strPathToTemp)
         assertTrue { tempDir.exists() }
     }
+
     override fun constructOutPath(directoryName: String) = ""
 }
